move function closer to where it is used
