+++ busy_beaver/blueprints/integration/slack/block_kit.py
@@ -1,6 +1,3 @@
-from typing import Optional
-
-
 class Block:
     """Generic block object, serves as a base class for Slack Block Components
 
@@ -11,7 +8,7 @@ class Block:
 
     type = None
 
-    def __init__(self, block_id=None):
+    def __init__(self, block_id: str = ""):
         self.output = {}
         if block_id:
             self.output["block_id"] = block_id
@@ -32,7 +29,7 @@ class Section(Block):
 
     type = "section"
 
-    def __init__(self, text: str, block_id: Optional[str] = None):
+    def __init__(self, text: str, block_id: str = ""):
         super().__init__(block_id)
         self.output["type"] = self.type
         self.output["text"] = {"type": "mrkdwn", "text": text}
@@ -48,7 +45,7 @@ class Context(Block):
 
     type = "context"
 
-    def __init__(self, text: str = "", block_id: Optional[str] = None):
+    def __init__(self, text: str = "", block_id: str = ""):
         super().__init__(block_id)
         self.output["type"] = self.type
         self.output["elements"] = [{"type": "plain_text", "emoji": True, "text": text}]
@@ -65,6 +62,6 @@ class Divider(Block):
 
     type = "divider"
 
-    def __init__(self, block_id: Optional[str] = None):
+    def __init__(self, block_id: str = ""):
         super().__init__(block_id)
         self.output["type"] = self.type
