+++ busy_beaver/toolbox/__init__.py
@@ -1,2 +1,2 @@
 from .event_emitter import EventEmitter  # noqa
-from .helpers import make_response, make_slack_response, utc_now_minus  # noqa
+from .helpers import make_response, utc_now_minus  # noqa
